# src/main/java/com/belong/phone/api/phone_number_swaggerV1.yml
swagger: "2.0"
info:
  description: ""
  version: "1.0.0"
  title: "Phone Service"
  contact:
    email: "ernarangsumit@gmail.com"
host: "localhost:8080"
basePath: "/v1"
tags:
- name: "phones"
  description: "Phone numbers"
schemes:
- "http"
paths:
  /phones:
    get:
      tags:
      - "phone"
      summary: "Get all the phones from database"
      description: ""
      operationId: "getPhones"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Phone"
        "500":
          description: "Internal Server Error"
  /customers/{customerId}/phones:
    get:
      tags:
      - "customer"
      summary: "Finds all phones for a customer"
      description: ""
      operationId: "getAllPhoneNumbersForCustomer"
      produces:
      - "application/json"
      parameters:
      - name: "customerId"
        in: "path"
        description: "ID of customer to return the phone numbers"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Customer"
        "400":
          description: "Invalid status value"
  /customers/{customerId}/phones/activate/:
    put:
      tags:
      - "customer"
      summary: "Activate customer phone number"
      description: ""
      operationId: "activatePhoneNumber"
      produces:
      - "application/json"
      parameters:
      - name: "customerId"
        in: "path"
        description: "ID of customer to return the phone numbers"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "Updated user object"
        required: true
        schema:
          $ref: "#/definitions/Customer"  
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Customer"
        "400":
          description: "Invalid customer id"
definitions:
  Phone:
    type: "object"
    required:
    - "number"
    - "status"
    properties:
      number:
        type: "integer"
        format: "int64"
      status:
        type: "string"
  Customer:
    type: "object"
    required:
    - "id"
    properties:
      id:
        type: "integer"
        format: "int64"
      phones:
        type: "array"
        items:
          $ref: "#/definitions/Phone"